fix calculation of milliseconds since epoch (int overflow)

+++ WeatherApplet.cxx
@@ -223,7 +223,7 @@ void WeatherApplet::gotData(KJob *job)
 
   if ( data.contains("weather") && data["weather"].isArray() )
   {
-    QDateTime dt = QDateTime::fromMSecsSinceEpoch(data["dt"].toInt() * 1000);
+    QDateTime dt = QDateTime::fromMSecsSinceEpoch(qint64(data["dt"].toInt()) * 1000);
     shortForecast[0]->day->setText(dt.time().toString(Qt::SystemLocaleShortDate));
     setIcon(shortForecast[0]->icon, data["weather"].toArray()[0].toObject()["icon"].toString());
   }
@@ -240,7 +240,7 @@ void WeatherApplet::gotData(KJob *job)
     for (int i = 0; i < 3; i++)
     {
       setIcon(shortForecast[1 + i]->icon, array[i].toObject()["weather"].toArray()[0].toObject()["icon"].toString());
-      QDateTime dt = QDateTime::fromMSecsSinceEpoch(array[i].toObject()["dt"].toInt() * 1000);
+      QDateTime dt = QDateTime::fromMSecsSinceEpoch(qint64(array[i].toObject()["dt"].toInt()) * 1000);
       shortForecast[1 + i]->day->setText(dt.time().toString(Qt::SystemLocaleShortDate));
       shortForecast[1 + i]->show();
     }
@@ -251,7 +251,7 @@ void WeatherApplet::gotData(KJob *job)
     {
       QJsonObject obj = value.toObject();
 
-      int day = QDateTime::fromMSecsSinceEpoch(obj["dt"].toInt() * 1000).date().dayOfWeek();
+      int day = QDateTime::fromMSecsSinceEpoch(qint64(obj["dt"].toInt()) * 1000).date().dayOfWeek();
       double temp = obj["main"].toObject()["temp"].toDouble();
 
       if ( !minTemp.contains(day) )
@@ -276,7 +276,7 @@ void WeatherApplet::gotData(KJob *job)
         QString icon = obj["weather"].toArray()[0].toObject()["icon"].toString();
         setIcon(forecast[idx]->icon, icon);
 
-        int day = QDateTime::fromMSecsSinceEpoch(obj["dt"].toInt() * 1000).date().dayOfWeek();
+        int day = QDateTime::fromMSecsSinceEpoch(qint64(obj["dt"].toInt()) * 1000).date().dayOfWeek();
         forecast[idx]->day->setText(QDate::shortDayName(day));
         forecast[idx]->min->setText(i18n("%1 %2", locale().toString(minTemp[day], 'f', 1), tempUnit));
         forecast[idx]->max->setText(i18n("%1 %2", locale().toString(maxTemp[day], 'f', 1), tempUnit));
